+++ NEWS.md
@@ -4,6 +4,7 @@
 -
 
 ### Fixed
+- ([#18718](https://github.com/rstudio/rstudio/issues/18718)): Hardened R option and version cache initialization against R errors leaving a C++ static-initialization guard locked during session startup.
 -
 
 ### Dependencies